+++ tests/seed/fake_data.ts
@@ -1,4 +1,17 @@
 import { v4 as uuidv4 } from 'uuid';
+import { buildBetaEmail } from '../../src/controllers/utils';
+import Betagouv from '../../src/betagouv';
+import knex from '../../src/db';
+
+const populatePrimaryEmail = async() => {
+  const allOvhEmails = await Betagouv.getAllEmailInfos();
+  for (const emailId of allOvhEmails) {
+      await knex('users').insert({
+          username: emailId,
+          primary_email: buildBetaEmail(emailId)
+      })
+  }
+}
 
 export async function seed(knex) {
   // A few universities
@@ -43,5 +56,9 @@ export async function seed(knex) {
   ];
 
   await knex("newsletters").insert(newsletterList);
+
   console.log(`inserted ${newsletterList.length} fake data to newsletters table`);
+
+  populatePrimaryEmail()
+  console.log('Populate users table with existing emails')
 };
